+++ src/Hoyo/Internal/Parse.hs
@@ -9,6 +9,8 @@ Parse CLI arguments.
 
 module Hoyo.Internal.Parse where
 
+import           Data.Char                       (isSpace)
+import           Data.List
 import qualified Data.Text                       as T
 
 import           Hoyo.CLI.Complete
@@ -19,6 +21,10 @@ import           Hoyo.Internal.Version
 import           Options.Applicative
 import           Options.Applicative.Help.Chunk
 import           Options.Applicative.Help.Pretty
+
+import           System.Environment
+import           System.Exit
+import           System.Pager
 
 import           Text.Read
 
@@ -145,6 +151,11 @@ listCommand = List <$> (
                           <> metavar "<directory>"
                           <> help "Search bookmarks by directory"
                         ))
+                    <*> switch (
+                          long "json"
+                          <> short 'j'
+                          <> help "List bookmarks in JSON format"
+                        )
                   )
 
 -- | Parse options for the @hoyo clear@ command.
@@ -175,7 +186,12 @@ configCommand = ConfigCmd <$> hsubparser (
 
 -- | Parse options for the @hoyo config print@ sub-command.
 configPrintCommand :: Parser ConfigCommand
-configPrintCommand = pure (Print ConfigPrintOptions)
+configPrintCommand = Print . ConfigPrintOptions
+                        <$> switch (
+                              long "json"
+                              <> short 'j'
+                              <> help "Print config in JSON format"
+                            )
 
 -- | Parse options for the @hoyo config reset@ sub-command.
 configResetCommand :: Parser ConfigCommand
@@ -297,11 +313,15 @@ options = info (parseOptions <**> helper) (
           )
 
 -- | Show the help message. Pass a non-'Nothing' argument to specify a command.
+-- If the help message is longer than a page, pass it through the system's pager.
 showHelp :: Maybe String -> IO ()
-showHelp cmd =
-  handleParseResult
-    $ Failure
-    $ parserFailure defaultPrefs options (ShowHelpText cmd) []
+showHelp cmd = do
+  let failure = parserFailure defaultPrefs options (ShowHelpText cmd) []
+  progn <- getProgName
+  let (msg, exit) = renderFailure failure progn
+  let msg' = dropWhileEnd isSpace msg <> "\n"
+  printOrPage $ T.pack msg'
+  exitWith exit
 
 -- | Split a string into arguments as they would be interpreted on the command line.
 --